Move implementation of API to internal
