+++ graphql-api.cabal
@@ -2,7 +2,7 @@
 --
 -- see: https://github.com/sol/hpack
 --
--- hash: 2160ff0226a0b10f7166ddbcf09ec6f79dfddfd66c81244898fb1afabe36bcf8
+-- hash: 988cb409d92080d6a3873f04afe11a0a68a3d3be6134dd00c7303620b081063e
 
 name:           graphql-api
 version:        0.3.0
@@ -52,6 +52,7 @@ library
       GraphQL
       GraphQL.API
       GraphQL.API.Enum
+      GraphQL.Internal.API
       GraphQL.Internal.Arbitrary
       GraphQL.Internal.Execution
       GraphQL.Internal.Name
